Add hostname resolution hook
This allows users to provide a form of host resolution by subclassing
FileAdapter and overriding the resolve_host() method.  Note that it only
allows for resolution back to an actual filename currently.

This can support things like automounting or relative pathnames, or
anything that requires some processing before resolving to a file.
