+++ setup.py
@@ -32,4 +32,5 @@ setup(
         'Programming Language :: Python :: 3.2',
         'Programming Language :: Python :: 3.3',
     ),
+    test_suite="tests"
 )
